Switched to Clerk for auth.

+++ src/app/user/d4h-access-keys/actions.ts
@@ -2,8 +2,10 @@
 
 import { revalidatePath } from 'next/cache'
 
-import { getUserSession } from '@/lib/get-user-session'
+import { currentUser } from '@clerk/nextjs/server'
+
 import prisma from '@/lib/prisma'
+
 
 
 
@@ -14,21 +16,22 @@ type CreateArgs =  { key: string, label: string, memberId: number, teamName: str
  */
 export async function createAccessKey({ key, label, memberId, teamName, teamId, primary }: CreateArgs) {
 
-    const { userId } = await getUserSession()
+    const user = await currentUser()
+    if(!user) throw new Error("Must be logged in to execute action 'createAccessKey'")
 
     if(primary) {
         // This new key will become the primary so we need to mark all of the old ones as not primary
         prisma.d4hAccessKey.updateMany({
-            where: { userId, primary: true },
+            where: { userId: user.id, primary: true },
             data: { primary: false }
         })
     }
 
     await prisma.d4hAccessKey.create({
-        data: { userId, key, label, memberId, teamName, teamId, primary, enabled: true }
+        data: { userId: user.id, key, label, memberId, teamName, teamId, primary, enabled: true }
     })
 
-    revalidatePath('/~/d4h-access-keys')
+    revalidatePath('/user/d4h-access-keys')
 }
 
 
@@ -38,22 +41,23 @@ type UpdateArgs = { id: string, label: string, primary: boolean, enabled: boolea
  * Server Action to update an access key record in the database.
  */
 export async function updateAccessKey({ id, label, primary, enabled }: UpdateArgs) {
-    const { userId } = await getUserSession()
+    const user = await currentUser()
+    if(!user) throw new Error("Must be logged in to execute action 'updateAccessKey'")
 
     if(primary) {
         // This key will become the primary so we need to mark all others as not primary
         prisma.d4hAccessKey.updateMany({
-            where: { userId, primary: true },
+            where: { userId: user.id, primary: true },
             data: { primary: false }
         })
     }
 
     await prisma.d4hAccessKey.update({
-        where: { id, userId },
+        where: { id, userId: user.id },
         data: { label, primary, enabled }
     })
 
-    revalidatePath('/~/d4h-access-keys')
+    revalidatePath('/user/d4h-access-keys')
 }
 
 
@@ -63,11 +67,12 @@ type DeleteArgs = { id: string }
  * Server Action to delete an access key record from the database.
  */
 export async function deleteAccessKey({ id }: DeleteArgs) {
-    const { userId } = await getUserSession()
+    const user = await currentUser()
+    if(!user) throw new Error("Must be logged in to execute action 'deleteAccessKey'")
 
     await prisma.d4hAccessKey.delete({
-        where: { id, userId }
+        where: { id, userId: user.id }
     })
 
-    revalidatePath('/~/d4h-access-keys')
+    revalidatePath('/user/d4h-access-keys')
 }
